+++ CHANGELOG.md
@@ -12,6 +12,14 @@ Emojis for the following are chosen based on [gitmoji](https://gitmoji.dev/).
 
 ## [Upcoming] i18n-check 1.x
 
+## i18n-check 1.14.0
+
+### ✨ Features
+
+- The CLI now sorts all required locale files if the `sorted-keys` check is activated when `--fix` options are used for the `invalid-keys`, `missing-keys` and `nonexistent-keys` checks ([#78](https://github.com/activist-org/i18n-check/issues/78)).
+- The `sorted-keys` check only functions if the `repeat-keys` check has passed for the given JSON file.
+  - The user is warned if there are repeated keys so these can be fixed before sorting.
+
 ## i18n-check 1.13.0
 
 ### ✨ Features